hr_hire_requisitions

+++ app/Http/Controllers/Web/HumanResource/Interview/InterviewQuestionController.php
@@ -1,25 +1,27 @@
 <?php
 namespace App\Http\Controllers\Web\HumanResource\Interview;
 
-use App\Exceptions\GeneralException;
 use Illuminate\Http\Request;
+use Illuminate\Support\Facades\DB;
+use App\Exceptions\GeneralException;
 use App\Http\Controllers\Controller;
 use App\Repositories\Access\UserRepository;
 use App\Models\HumanResource\Interview\Question;
 use App\Repositories\Unit\DesignationRepository;
 use App\Models\HumanResource\Interview\Interview;
-use App\Models\HumanResource\Interview\InterviewApplicantMarks;
+use App\Models\HumanResource\Interview\InterviewTypes;
 use App\Models\HumanResource\Interview\InterviewPanelist;
 use App\Models\HumanResource\Interview\InterviewQuestion;
+use App\Models\HumanResource\Interview\InterviewPanelistMarks;
+use App\Models\HumanResource\Interview\InterviewApplicantMarks;
+use App\Models\HumanResource\Interview\InterviewPanelistCounter;
+use App\Models\HumanResource\Interview\InterviewSchedule;
 use App\Repositories\HumanResource\Interview\InterviewRepository;
 use App\Repositories\HumanResource\Interview\InterviewQuestionRepository;
 use App\Repositories\HumanResource\Interview\InterviewApplicantRepository;
 use App\Repositories\HumanResource\HireRequisition\HrHireApplicantRepository;
 use App\Repositories\HumanResource\Interview\InterviewQuestionMarksRepository;
 use App\Repositories\HumanResource\HireRequisition\HireRequisitionJobRepository;
-use Illuminate\Support\Facades\DB;
-use App\Models\HumanResource\Interview\InterviewPanelistMarks;
-use App\Models\HumanResource\Interview\InterviewPanelistCounter;
 
 class InterviewQuestionController extends Controller
 {
@@ -52,7 +54,8 @@ class InterviewQuestionController extends Controller
 
         $total_questions_marks = $this->interviewQuestionRepository->query()->where('interview_id',$request->interview_id)->sum('marks');
         $marks = $total_questions_marks + $request->marks;
-        if($total_questions_marks > 100 ){
+       
+        if($marks > 100 ){
             throw new GeneralException('Total Marks Cannot Exceed 100');
         }
         $this->interviewQuestionRepository->store($request->all());      
@@ -70,11 +73,17 @@ class InterviewQuestionController extends Controller
     }
 
     public function create(Interview $interview){
-        $questions = $this->interviewQuestionRepository
-                    ->query()
-                    ->where('interview_id',$interview->id)
-                    ->get();
+        $questions = $this->interviewQuestionRepository->query()->where('interview_id',$interview->id)->get();
+        $hrHireRequisitionJob = $this->hireRequisitionJobRepository->query()->with('designation')->where('hr_hire_requisitions_jobs.id', $interview->hr_requisition_job_id)
+                                ->first();                
+        $interview_type = InterviewTypes::find($interview->interview_type_id);
+        $interview_date = InterviewSchedule::select(DB::raw("DATE(interview_date) as interview_date"))->where('interview_id',$interview->id)->get()->implode('interview_date',',');
+        $job_title = $this->designationRepository->getQueryDesignationUnit()->where('designations.id', $hrHireRequisitionJob->designation_id)
+                    ->first(); 
         return view('HumanResource.Interview.question')
+                ->with('job_title',$job_title)
+                ->with('interview_type',$interview_type)
+                ->with('interview_date',$interview_date)
                 ->with('questions',$questions)
                 ->with('interview',$interview);
     }
@@ -87,12 +96,24 @@ class InterviewQuestionController extends Controller
 
 
     public function addQuestionMarks(Interview $interview){
+        $interview_type = InterviewTypes::find($interview->interview_type_id);;
+
+        $hrHireRequisitionJob = $this->hireRequisitionJobRepository
+            ->query()
+            ->with('designation')
+            ->where('hr_hire_requisitions_jobs.id', $interview->hr_requisition_job_id)
+            ->first();
+        $job_title = $this->designationRepository->getQueryDesignationUnit()
+            ->where('designations.id', $hrHireRequisitionJob->designation_id)
+            ->first();
         $questions = $this->interviewQuestionRepository
                     ->query()
                     ->where('interview_id',$interview->id)
                     ->get();
         return view('HumanResource.Interview.question_marks')
                 ->with('questions',$questions)
+                ->with('job_title',$job_title)
+                ->with('interview_type',$interview_type)
                 ->with('interview',$interview);
     }
     public function storeMarks(Request $request){
@@ -104,11 +125,7 @@ class InterviewQuestionController extends Controller
         $input['applicant_id'] = $request->applicant_id;
         try{
             DB::beginTransaction();
-                $this->interviewQuestionMarksRepository
-                        ->query()
-                        ->where('interview_id',$interview_id)
-                        ->where('applicant_id',$applicant_id)
-                        ->delete();
+                $this->interviewQuestionMarksRepository->query()->where('interview_id',$interview_id)->where('applicant_id',$applicant_id)->delete();
                 $total_marks = 0;
                 for($i=1; $i<=$total_questions; $i++){
                     $question = $request->input('question'.$i);
@@ -120,27 +137,24 @@ class InterviewQuestionController extends Controller
                     $this->interviewQuestionMarksRepository->store($input);
                     $total_marks +=  $marks;
                 }
-                InterviewPanelistMarks::where('interview_id', $interview_id)
-                                                ->where('panelist_id', access()->id())
-                                                ->where('applicant_id', $applicant_id)->delete();
+                InterviewPanelistMarks::where('interview_id', $interview_id)->where('panelist_id', access()->id())->where('applicant_id', $applicant_id)->delete();                             
                 InterviewPanelistMarks::create([
                     'interview_id'=> $interview_id,
                     'panelist_id'=> access()->id(),
                     'marks'=> $total_marks,
                     'applicant_id' => $applicant_id,
                 ]);
-                $total_panelist_marks = InterviewPanelistMarks::where('interview_id', $interview_id)
-                                                                ->where('applicant_id', $applicant_id)->sum('marks');
-               
-                $interviewApplicantMarks = InterviewApplicantMarks::where('interview_id', $interview_id)
-                                          ->where('applicant_id',$applicant_id)
-                                          ->first();
+                $total_panelist_marks = InterviewPanelistMarks::where('interview_id', $interview_id)->where('applicant_id', $applicant_id)->sum('marks');
+                // $interviewApplicantMarks = InterviewApplicantMarks::where('interview_id', $interview_id)->where('applicant_id',$applicant_id)->delete();
+                // return $total_panelist_marks;
+                $interviewApplicantMarks = InterviewApplicantMarks::where('interview_id', $interview_id)->where('applicant_id',$applicant_id)->first();
                 if(!is_null($interviewApplicantMarks)){
                     $total = ($total_panelist_marks / $total_panelist->total_panelist);
-                    $interviewApplicantMarks->update(['marks'=>$total]);                    
+                    $interviewApplicantMarks->update(['marks'=>$total]);   
                 }else{
+                    $total = ($total_panelist_marks / $total_panelist->total_panelist);
                     InterviewApplicantMarks::create(['interview_id'=> $interview_id,
-                    'marks'=> $total_panelist_marks,
+                    'marks'=> $total,
                     'applicant_id' => $applicant_id]);
                 }              
                 $this->interviewApplicantRepository->query()
